refactor: Angular pure functions to injectable services
