+++ libs/data-access/src/index.ts
@@ -9,5 +9,33 @@ export * from './lib/services/workbook/workbook.service';
 export * from './lib/services/xlsx-display/xlsx-display.service';
 export * from './lib/services/xlsx-facade/xlsx-facade.service';
 export * from './lib/services/xlsx-process/xlsx-process.service';
-export * from './lib/services/converter/utilities/utilities';
+export * from './lib/services/converter/utils/utils.service';
 export * from './lib/app.worker';
+
+// converter services
+// managers
+export * from './lib/services/converter/gtm-json-manager/managers/tag-manager.service';
+export * from './lib/services/converter/gtm-json-manager/managers/trigger-manager.service';
+export * from './lib/services/converter/gtm-json-manager/managers/variable-manager.service';
+
+// tags
+export * from './lib/services/converter/gtm-json-manager/tags/event-tag.service';
+export * from './lib/services/converter/gtm-json-manager/tags/google-tag.service';
+export * from './lib/services/converter/gtm-json-manager/tags/scroll-tag.service';
+export * from './lib/services/converter/gtm-json-manager/tags/video-tag.service';
+
+// triggers
+export * from './lib/services/converter/gtm-json-manager/triggers/event-trigger.service';
+export * from './lib/services/converter/gtm-json-manager/triggers/scroll-trigger.service';
+export * from './lib/services/converter/gtm-json-manager/triggers/video-trigger.service';
+export * from './lib/services/converter/gtm-json-manager/triggers/trigger-utils.service';
+
+// variables
+export * from './lib/services/converter/gtm-json-manager/variables/custom-variable.service';
+export * from './lib/services/converter/gtm-json-manager/variables/data-layer-variable.service';
+export * from './lib/services/converter/gtm-json-manager/variables/regex-variable.service';
+export * from './lib/services/converter/gtm-json-manager/variables/scroll-variable.service';
+export * from './lib/services/converter/gtm-json-manager/variables/video-variable.service';
+
+// utilities
+export * from './lib/services/converter/gtm-json-manager/parameter-utils.service';
