+++ libs/data-access/src/lib/services/converter/gtm-json-manager/tags/tags.spec.ts
@@ -4,11 +4,10 @@ jest.mock('../utilities/utilities', () => ({
   hasExistedDataLayer: jest.fn().mockReturnValue(false),
 }));
 
-import { Tag, TriggerConfig } from 'src/app/interfaces/gtm-config-generator';
+import { Tag, TriggerConfig } from '@utils';
 import { createTag } from './event-tag';
 import { createGA4Configuration } from './ga4-configuration-tag';
 import { createScrollTag } from './scroll-tag';
-import { getTags } from './tag-utilities';
 import { createVideoTag } from './video-tag';
 
 const mockTags: Tag[] = [
