--- libs/data-access/src/lib/services/converter/gtm-json-manager/triggers/event-trigger.ts
@@ -1,24 +1,0 @@
-import { TriggerConfig } from '@utils';
-import { createTemplateParameter } from '../parameter-utils';
-
-export function createTrigger(
-  accountId: string,
-  containerId: string,
-  trigger: string
-): TriggerConfig {
-  return {
-    accountId,
-    containerId,
-    type: 'CUSTOM_EVENT',
-    name: `event equals ${trigger}`,
-    customEventFilter: [
-      {
-        type: 'EQUALS',
-        parameter: [
-          createTemplateParameter('arg0', '{{_event}}'),
-          createTemplateParameter('arg1', trigger),
-        ],
-      },
-    ],
-  };
-}
