--- libs/data-access/src/lib/services/converter/gtm-json-manager/triggers/trigger-utilities.ts
@@ -1,22 +1,0 @@
-import { Trigger, TriggerConfig } from '@utils';
-import { createVideoTrigger } from './video-trigger';
-import { createScrollTrigger } from './scroll-trigger';
-import { createTrigger } from './event-trigger';
-
-export function getTriggers(
-  accountId: string,
-  containerId: string,
-  data: Record<string, string>[],
-  triggers: Trigger[]
-): TriggerConfig[] {
-  return [
-    ...triggers.map(({ name: trigger }) => {
-      return createTrigger(accountId, containerId, trigger);
-    }),
-    ...createVideoTrigger(accountId, containerId, data),
-    ...createScrollTrigger(accountId, containerId, data),
-  ].map((_trigger, index) => ({
-    ..._trigger,
-    triggerId: (index + 1).toString(),
-  }));
-}
