--- libs/data-access/src/lib/services/converter/gtm-json-manager/variables/data-layer-variable.ts
@@ -1,24 +1,0 @@
-import { VariableConfig } from '@utils';
-import {
-  createBooleanParameter,
-  createIntegerParameter,
-  createTemplateParameter,
-} from '../parameter-utils';
-
-export function createVariable(
-  accountId: string,
-  containerId: string,
-  dataLayerName: string
-): VariableConfig {
-  return {
-    name: `DLV - ${dataLayerName.trim()}`,
-    type: 'v',
-    accountId,
-    containerId,
-    parameter: [
-      createIntegerParameter('dataLayerVersion', '2'),
-      createBooleanParameter('setDefaultValue', 'false'),
-      createTemplateParameter('name', dataLayerName.trim()),
-    ],
-  };
-}
