--- libs/data-access/src/lib/services/converter/gtm-json-manager/variables/regex-measurement-id-variable.ts
@@ -1,27 +1,0 @@
-import { VariableConfig } from '@utils';
-import {
-  createBooleanParameter,
-  createTemplateParameter,
-} from '../parameter-utils';
-
-export function createRegexMeasurementIdVariable(
-  accountId: string,
-  containerId: string
-): VariableConfig {
-  return {
-    name: `Measurement ID`,
-    type: 'remm',
-    accountId,
-    containerId,
-    parameter: [
-      createBooleanParameter('setDefaultValue', 'true'),
-      createTemplateParameter('input', '{{Page URL}}'),
-      createBooleanParameter('fullMatch', 'false'),
-      createBooleanParameter('replaceAfterMatch', 'false'),
-      createTemplateParameter('defaultValue', 'G-1'),
-      createBooleanParameter('ignoreCase', 'true'),
-    ],
-    fingerprint: '1696861232768',
-    formatValue: {},
-  };
-}
