--- libs/data-access/src/lib/services/converter/gtm-json-manager/variables/scroll-variable.ts
@@ -1,18 +1,0 @@
-import { VariableConfig } from '@utils';
-
-export function scrollBuiltInVariable({
-  accountId,
-  containerId,
-}: {
-  accountId: string;
-  containerId: string;
-}): VariableConfig[] {
-  return [
-    {
-      accountId,
-      containerId,
-      type: 'SCROLL_DEPTH_THRESHOLD',
-      name: 'Scroll Depth Threshold',
-    },
-  ];
-}
