+++ libs/data-access/src/lib/services/converter/utils/data-layer-utils.service.ts
@@ -1,3 +1,8 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root',
+})
 export class DataLayerManager {
   dataLayers: string[] = [];
