+++ libs/data-access/src/lib/services/xlsx-facade/xlsx-helper.spec.ts
@@ -1,8 +1,3 @@
-import {
-  filterGtmSpecsFromData,
-  convertSpecStringToObject,
-} from './xlsx-helper';
-
 describe('Helper Functions', () => {
   describe('filterGtmSpecsFromData', () => {
     it.each([
